occupancy_grid_display: fix deadlocks on resubscribe
Because the tf message filter holds its lock during the callback,
it is not legal to do anything that will grab the tf lock while
holding our mutex, as possible deadlocks can occur.

Move resubscribing onto a one-shot timer, and be very careful to
not hold our lock when manipulating the subscription filters.