add rviz TF API compatibility layer
In RViz 1.14 support for getTFClient was dropped. However, this
plugin needs to remain compatible with older versions of RViz prior
to getTF2BufferPtr existing in order to be used with kinetic. Provide
a compatilbity layer which detects whether getTF2BufferPtr is
declared using decltype and std::enable_if.  When present, this layer
uses the getTF2BufferPtr, and if absent, goes back to getTFClient.
Once this plugin no longer needs to support older versions of RViz,
this commit can be reverted.